se implemntas modulos
