--- component/App.js
@@ -1,8 +1,0 @@
-// const readline = require("readline");
-
-// const rl = readline.createInterface({
-//     input: process.stdin,
-//     output: process.stdout,
-//   });
-
-//   module.exports= rl;
