--- component/verTarea.js
@@ -1,13 +1,0 @@
-// const promptMenu = require('./mostrarMenu')
-
-  
-// const taskList = () => {
-//     console.log("las tareas almacenadas son: ");
-//     for (const task of tasks) {
-//       console.log(`id ${task.id} tarea ${task.description}`);
-//     }
-//   };
-
-//   module.exports = 
-//     taskList
-
